moved Explorer to pages and added conditional rendering
